Fix infinitely calling filtering method
