+++ frontend/src/app/page.js
@@ -1,9 +1,10 @@
-import { DataView } from "@/app/components/data-view";
+// import { DataView } from "@/app/components/data-view";
+import DashboardContainer from "@/app/components/dashboard-container";
 
 export default function Home() {
   return (
     <main>
-      <DataView path="home" />
+      <DashboardContainer path="home" />
     </main>
   );
 }
